xyzFieldControl done
moved old coilControll
